move out to selection, simplify if expressions
