+++ packages/@react-aria/grid/src/useGrid.ts
@@ -22,7 +22,7 @@ import {HTMLAttributes, Key, RefObject, useMemo, useRef} from 'react';
 import intlMessages from '../intl/*.json';
 import {useCollator, useLocale, useMessageFormatter} from '@react-aria/i18n';
 import {useInteractionModality} from '@react-aria/interactions';
-import {useSelectableCollection} from '@react-aria/selection';
+import {useHighlightSelectionDescription, useSelectableCollection} from '@react-aria/selection';
 
 export interface GridProps extends DOMProps, AriaLabelingProps {
   /** Whether the grid uses virtual scrolling. */
@@ -98,28 +98,7 @@ export function useGrid<T>(props: GridProps, state: GridState<T, GridCollection<
   let id = useId();
   gridKeyboardDelegates.set(state, delegate);
 
-  // do we have something that might be better suited for this? issue is that useInteractionModality
-  // doesn't return quite enough information
-  // if is on touch start is in the window put the message?
-  // grab from dnd?
-  // navigator touch points?
-  let modality = useInteractionModality();
-  let shouldLongPress = (modality === 'pointer' || modality === 'virtual') && 'ontouchstart' in window;
-
-  let interactionDescription = useMemo(() => {
-    let selectionMode = state.selectionManager.selectionMode;
-    let selectionBehavior = state.selectionManager.selectionBehavior;
-    // if we're in replace but can select multiple, then when using touch it's long press to enter selection mode
-    // if we can't tell what mode we're in, then we're probably in voice over, in which case we need the description to be set
-    // before the user interacts so they know they can enter selection mode before they actually have to interact
-    let message = undefined;
-    if (shouldLongPress) {
-      message = formatMessage('longPressToSelect');
-    }
-    return selectionBehavior === 'replace' && selectionMode !== 'none' && state.selectionManager.hasItemActions ? message : undefined;
-  }, [state.selectionManager.selectionMode, state.selectionManager.selectionBehavior, formatMessage, shouldLongPress]);
-
-  let descriptionProps = useDescription(interactionDescription);
+  let descriptionProps = useHighlightSelectionDescription(props, state);
 
   let domProps = filterDOMProps(props, {labelable: true});
   let gridProps: HTMLAttributes<HTMLElement> = mergeProps(
@@ -144,8 +123,8 @@ export function useGrid<T>(props: GridProps, state: GridState<T, GridCollection<
   let lastSelection = useRef(selection);
   useUpdateEffect(() => {
     if (!state.selectionManager.isFocused) {
-      // do we still need this? can't recall why i removed it, will need to see why it was added and add a test for whichever way it's supposed to be
-      // lastSelection.current = selection;
+      lastSelection.current = selection;
+
       return;
     }
 
@@ -155,23 +134,24 @@ export function useGrid<T>(props: GridProps, state: GridState<T, GridCollection<
     // If adding or removing a single row from the selection, announce the name of that item.
     let isReplace = state.selectionManager.selectionBehavior === 'replace';
     let messages = [];
-    if (addedKeys.size === 1 && (removedKeys.size === 0 || (state.selectionManager.selectedKeys.size === 1 && isReplace))) {
-      let addedText = getRowText(addedKeys.keys().next().value);
-      if (addedText) {
-        messages.push(formatMessage('selectedItem', {item: addedText}));
-      }
-    } else if (removedKeys.size === 1 && (addedKeys.size === 0 || (state.selectionManager.selectedKeys.size === 1 && isReplace))) {
-      if (state.collection.getItem(removedKeys.keys().next().value)) {
-        let removedText = getRowText(removedKeys.keys().next().value);
-        if (removedText) {
-          messages.push(formatMessage('deselectedItem', {item: removedText}));
-        }
-      }
-    } else if (isReplace && (lastSelection.current === 'all' || removedKeys.size > 0) && state.selectionManager.selectedKeys.size === 1) {
+
+    if ((state.selectionManager.selectedKeys.size === 1 && isReplace)) {
       if (state.collection.getItem(state.selectionManager.selectedKeys.keys().next().value)) {
         let currentSelectionText = getRowText(state.selectionManager.selectedKeys.keys().next().value);
         if (currentSelectionText) {
           messages.push(formatMessage('selectedItem', {item: currentSelectionText}));
+        }
+      }
+    } else if (addedKeys.size === 1 && removedKeys.size === 0) {
+      let addedText = getRowText(addedKeys.keys().next().value);
+      if (addedText) {
+        messages.push(formatMessage('selectedItem', {item: addedText}));
+      }
+    } else if (removedKeys.size === 1 && addedKeys.size === 0) {
+      if (state.collection.getItem(removedKeys.keys().next().value)) {
+        let removedText = getRowText(removedKeys.keys().next().value);
+        if (removedText) {
+          messages.push(formatMessage('deselectedItem', {item: removedText}));
         }
       }
     }
