+++ packages/@react-aria/selection/intl/ar-AE.json
@@ -1,0 +1,3 @@
+{
+  "longPressToSelect": "Long press to enter selection mode."
+}
