+++ packages/@react-aria/selection/intl/en-US.json
@@ -1,0 +1,3 @@
+{
+  "longPressToSelect": "Long press to enter selection mode."
+}
